LPS-64316 Don't use ServiceTracker in tests
https://issues.liferay.com/browse/LPS-64316
