+++ portal-test-integration/src/com/liferay/portal/search/test/BaseSearchResultUtilTestCase.java
@@ -18,16 +18,16 @@ import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
 import com.liferay.asset.kernel.model.AssetRenderer;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.portal.kernel.comment.Comment;
+import com.liferay.portal.kernel.model.ClassName;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.IndexerRegistry;
 import com.liferay.portal.kernel.search.RelatedSearchResult;
 import com.liferay.portal.kernel.search.SearchResult;
 import com.liferay.portal.kernel.search.result.SearchResultTranslator;
+import com.liferay.portal.kernel.service.ClassNameLocalService;
 import com.liferay.portal.kernel.util.FastDateFormatFactory;
 import com.liferay.portal.kernel.util.FastDateFormatFactoryUtil;
-import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.Props;
 import com.liferay.portal.kernel.util.PropsUtil;
 import com.liferay.registry.BasicRegistryImpl;
@@ -56,9 +56,9 @@ public abstract class BaseSearchResultUtilTestCase extends PowerMockito {
 
 		setUpRegistryUtil();
 
+		setUpClassNameLocalService();
 		setUpFastDateFormatFactoryUtil();
 		setUpIndexerRegistry();
-		setUpPortalUtil();
 		setUpPropsUtil();
 		setUpSearchResultTranslator();
 	}
@@ -98,6 +98,23 @@ public abstract class BaseSearchResultUtilTestCase extends PowerMockito {
 
 	protected abstract SearchResultTranslator createSearchResultTranslator();
 
+	protected void setUpClassNameLocalService() throws Exception {
+		ClassName className = Mockito.mock(ClassName.class);
+
+		when(
+			classNameLocalService.getClassName(
+				SearchTestUtil.ATTACHMENT_OWNER_CLASS_NAME_ID)
+		).thenReturn(
+			className
+		);
+
+		when(
+			className.getClassName()
+		).thenReturn(
+			SearchTestUtil.ATTACHMENT_OWNER_CLASS_NAME
+		);
+	}
+
 	protected void setUpFastDateFormatFactoryUtil() {
 		FastDateFormatFactoryUtil fastDateFormatFactoryUtil =
 			new FastDateFormatFactoryUtil();
@@ -111,20 +128,6 @@ public abstract class BaseSearchResultUtilTestCase extends PowerMockito {
 
 		registry.registerService(
 			IndexerRegistry.class, new TestIndexerRegistry());
-	}
-
-	protected void setUpPortalUtil() {
-		Portal portal = Mockito.mock(Portal.class);
-
-		when(
-			portal.getClassName(SearchTestUtil.ATTACHMENT_OWNER_CLASS_NAME_ID)
-		).thenReturn(
-			SearchTestUtil.ATTACHMENT_OWNER_CLASS_NAME
-		);
-
-		PortalUtil portalUtil = new PortalUtil();
-
-		portalUtil.setPortal(portal);
 	}
 
 	protected void setUpPropsUtil() {
@@ -148,6 +151,9 @@ public abstract class BaseSearchResultUtilTestCase extends PowerMockito {
 
 	@Mock
 	protected AssetRendererFactory<?> assetRendererFactory;
+
+	@Mock
+	protected ClassNameLocalService classNameLocalService;
 
 	protected SearchResultTranslator searchResultTranslator;
